### 2022-07-18 #4579 @editorialbot
[PRE REVIEW]: SpmImage Tycoon: Organize and analyze scanning probe microscopy data
**Submitting author:** @alexriss (<a href="http://orcid.org/0000-0002-3212-7925">Alexander Riss</a>)
**Repository:** https://github.com/alexriss/SpmImageTycoon.jl
**Branch with paper.md** (empty if default branch):
**Version:** 0.3.6
