### 2022-07-25 #4579 @Kevin-Mattheus-Moerman
@alexriss thanks for this submission. I am the AEiC on call this week. Can you check those potentially invalid DOI's? :point_up: Once you've ammended them you can call `@editorialbot check references` here and use `@editorialbot generate pdf` to update the paper.
