+++ JOSS/paper.bib
@@ -76,7 +76,7 @@ volume = {5},
 pages = {31-36},
 year = {2016},
 issn = {2352-7110},
-doi = {https://doi.org/10.1016/j.softx.2016.04.001},
+doi = {10.1016/j.softx.2016.04.001},
 url = {https://www.sciencedirect.com/science/article/pii/S2352711016300012},
 author = {Michael Ruby},
 }
