+++ JOSS/paper.bib
@@ -46,7 +46,7 @@
 
 @article{gwyddion,
 author = {David Nečas and Petr Klapetek},
-doi = {doi:10.2478/s11534-011-0096-2},
+doi = {10.2478/s11534-011-0096-2},
 url = {https://doi.org/10.2478/s11534-011-0096-2},
 title = {Gwyddion: an open-source software for SPM data analysis},
 journal = {Open Physics},
