### 2022-07-25 3a92d3a
ignore JOSS subfolder for CI
