+++ .github/workflows/CI.yml
@@ -5,11 +5,13 @@ on:
       - main
     paths-ignore:
       - '**.md'
+      - 'JOSS/**'
   pull_request:
     branches:
       - main
     paths-ignore:
       - '**.md'
+      - 'JOSS/**'
 jobs:
   test:
     name: Julia ${{ matrix.version }} - ${{ matrix.os }} - ${{ matrix.arch }} - ${{ github.event_name }}
