+++ Project.toml
@@ -41,7 +41,7 @@ NaturalSort = "^1.0"
 SpmImages = "^0.2"
 SpmSpectroscopy = "^0.5"
 StatsBase = "^0.33"
-ZipFile = "^0.9"
+ZipFile = "^0.9, 0.10"
 julia = "^1.7"
 
 [extras]
