### 2022-08-01 #4579 @jarvist
I don't have much background in scanning microscopy, unfortunately.
