### 2022-08-03 #4579 @jgostick
Hi @jingpengw and @xaviercm94, thanks for agreeing to review this submission.  A few things to note about JOSS reviews:

1. They are done publicly, so the authors and reviewers work together to improve the submission
2. The general aim is to help submissions improve until they are ready for publications rather than to reach an accept/reject decision
3. There will be a checklist of items for you to consider on the review thread, so your work is well defined.
4. We generally like to get it all done in about a month or less. Obviously, this is flexible if needed.
