### 2022-08-22 #4644 @kasasxav
Hi @alexriss,

I am still going through the review process, looks good so far! But there are a few requirements that you are missing:

In the documentation,

- Statement of need: you do write that SpmImage Tycoon is a "Cross-platform app to manage and edit scanning probe microscopy (SPM) images and spectra", but it would be nice if you provide more information about what problems it can solve or features it has. I know you have included that in the paper, but I think it would help if you have something similar in Github too.

- Community guidelines: there is no information about what are the guidelines for people that want to contribute to your software.

- I also miss documentation regarding the tests, how to run them and what do they cover?

- I see the scripts are commented, but I can't find a general documentation of the API. I am not sure if this is necessary so I leave it as a suggestion.

Regarding the paper,

- Statement of the field is missing, what are the available software packages in the field of scanning probe microscopy image analysis, and how does your solution compare to them?

- Just two small typos:
              line 11 -> into a OpenDocument -> into an OpenDocument
              line 34 -> stored in a HDF5 -> stored in an HDF5

- Would be nice to develop more in some areas:
             line 18-19: "However, much less emphasis has been placed on improving the managing and organization of the measured datasets".  Here you can develop in which work has been done.
             line 37: "Due to its modular design, new features can be simply added". How can one add new features?

- And one suggestion. You mention the included features in the summary, but you could also explain more them more in detail so the reader can fully capture the potential of your software. Maybe you could even include a figure.
