### 2022-08-22 #4644 @kasasxav
Hi again,

I am going through the videos and I installed the software, I was wondering if there is a dataset I could use for trying out the functionality?
Maybe it would be nice to provide it in general for the users, but otherwise do you know of any publicly available data?
