### 2022-08-22 #4644 @alexriss
HI - thank you very much for your efforts. I will go through your comments soon. I just wanted to mention that the app is geared towards "end users", i.e. people interacting via the GUI. Thus, I prefer to keep the docs simple. This is the reason that there is no documentation on the API. However, the libraries used by this app are released separately (SpmImages.jl, SpmSpectroscopy.jl) and more information can be found on the respective github repositories.

Regarding the tests, you can run them via `julia> ]` (to change into the pkg manager and then `pkg> test SpmImageTycoon`. They cover most function of the software and are run automatically via github actions. Again, I would prefer to not specifically mention this procedure in the docs because it should not be necessary for the end users to run these tests (they might even be overwhelmed by the instructions). And the more "julia-savy" probably already know how to run them if needed (this is pretty much standard practice for julia packages).

But if required, I can of course add this information.

There is a test dataset here:
https://github.com/alexriss/SpmImageTycoon.jl/tree/main/test/data
It is used for the automated tests., too.

Thanks again!
