### 2022-08-23 alexriss/SpmImageTycoon.jl#11 @kasasxav
Error or warning when opening directory test data
Hi!

I just got this error when opening the test data in Tycoon. If I click ok everything still works though.
