### 2022-08-23 #4644 @kasasxav
Hi @alexriss,

Thank you! Thank you for the explanations and test data, looks good then!

I also tried the software and followed the videos in youtube to check that the functionality was working, and everything was fine. The software presented, as I understand, is for users of SPM and similar techniques that wish to browse and analyze their images. There is a list of shortcuts easily accessible that I count as documentation, and one can easily browse and adjust the image contrast, trace line profiles to the data, etc. The software is very intuitive and easy to use, I think it would make an impact and help users navigate and organize their data. Furthermore, the images contain metadata that one can access, filter and copy.

My only concern is, if a developer wants to implement let's say an image processing algorithm into your software, how would they do it? or add new fields into the metadata? I think that would be useful to implement/document and facilitate the project to grow. For example, performing some kind of transformation to the data or filtering (smoothing or others).  I understand though that's not the focus of your project since it's targetted to end users, but not having certain functionality could prevent users to use the software. Maybe some users have expertise to contribute or other people in their groups that can help out with programming. I am not so familiar with SPM though so I don't know if that would be a common situation, but I could imagine this happening in other microscopy modalities. I am also not that familiar with Julia so I don't actually know if implementing new functionality would be hard. Could you maybe list the steps that would be required for that?

Thank you!

/Xavier.
