### 2022-08-23 #4644 @kasasxav
P.S. I would suggest to have the link to the metadata in the main repo or a possibility to browse through it from the main software. Then users can see the potential of the software even if they don't have available images
