### 2022-08-23 alexriss/SpmImageTycoon.jl#11 @alexriss
Thanks for the report. This error occurs when opening the directory? First open? Or re-open? It does not happen on my system. Can you paste the whole error stack from the console window? Also, which OS and julia version do you use?

I somehow suspect it is related to a third-party package I am using (I only use the `convert` function once - for conversion of old database formats, and it seems very unlikely that this was needed in your case). So, you can also try updating all packages in julia: press `julia> ]` to change into package manager, then type `pkg> up`.
