### 2022-08-23 b78b6fe
use julia version 1.8 for CI
