### 2022-08-23 alexriss/SpmImageTycoon.jl#11 @kasasxav
and it occurs when I open the directory, already the first time I do it. But also the following ones.
