### 2022-08-23 alexriss/SpmImageTycoon.jl#11 @kasasxav
I tried updating the packages but the error persists
