### 2022-08-23 563703d
initialize last_directories as array of String
