### 2022-08-23 alexriss/SpmImageTycoon.jl#11 @alexriss
I could not reproduce the error on windows, nor on Linux/WSL. However, I think I know where it came from, and I have released an update. You can update all packages as explained above - this time it should update `SpmImageTycoon` to version 0.3.7. Then restart julia and run the app (you can double-check the version in the app by clicking "About").
Please see if this fixes the error. Thanks!
