### 2022-08-23 aeccf66
add info on test data, expand technical background and how to contribute
