+++ CONTRIBUTING.md
@@ -1,0 +1,17 @@
+# Contributing
+
+Welcome and thank you for contributing to SpmImageTycoon.jl.
+
+Please open an issue if you:
+
+- want to report a bug
+- have problems using the package
+- have ideas for new features or ways to improve the usage of this package
+
+You can contribute (e.g. fix bugs, add new features) to this package via pull requests:
+
+- please first discuss your contributions in a new issue
+- ensure that all tests pass locally before starting the pull request
+- new features should be included in the tests
+
+Please follow our [Code of Conduct](CODE_OF_CONDUCT.md).
