+++ JOSS/paper.bib
@@ -73,10 +73,52 @@ URL = {https://doi.org/10.1063/1.2432410}
 title = {SpectraFox: A free open-source data management and analysis tool for scanning probe microscopy and spectroscopy},
 journal = {SoftwareX},
 volume = {5},
-pages = {31-36},
+pages = {31--36},
 year = {2016},
 issn = {2352-7110},
 doi = {10.1016/j.softx.2016.04.001},
 url = {https://www.sciencedirect.com/science/article/pii/S2352711016300012},
 author = {Michael Ruby},
 }
+
+@article{imagej,
+  doi = {10.1038/nmeth.2089},
+  url = {https://doi.org/10.1038/nmeth.2089},
+  year = {2012},
+  month = jun,
+  publisher = {Springer Science and Business Media {LLC}},
+  volume = {9},
+  number = {7},
+  pages = {671--675},
+  author = {Caroline A Schneider and Wayne S Rasband and Kevin W Eliceiri},
+  title = {{NIH} Image to {ImageJ}: 25 years of image analysis},
+  journal = {Nature Methods}
+}
+
+@article{imagej2,
+  doi = {10.1186/s12859-017-1934-z},
+  url = {https://doi.org/10.1186/s12859-017-1934-z},
+  year = {2017},
+  month = nov,
+  publisher = {Springer Science and Business Media {LLC}},
+  volume = {18},
+  number = {1},
+  author = {Curtis T. Rueden and Johannes Schindelin and Mark C. Hiner and Barry E. DeZonia and Alison E. Walter and Ellen T. Arena and Kevin W. Eliceiri},
+  title = {{ImageJ}2: {ImageJ} for the next generation of scientific image data},
+  journal = {{BMC} Bioinformatics}
+}
+
+@article{fiji,
+  doi = {10.1038/nmeth.2019},
+  url = {https://doi.org/10.1038/nmeth.2019},
+  year = {2012},
+  month = jun,
+  publisher = {Springer Science and Business Media {LLC}},
+  volume = {9},
+  number = {7},
+  pages = {676--682},
+  author = {Johannes Schindelin and Ignacio Arganda-Carreras and Erwin Frise and Verena Kaynig and Mark Longair and Tobias Pietzsch and Stephan Preibisch and Curtis Rueden and Stephan Saalfeld and Benjamin Schmid and Jean-Yves Tinevez and Daniel James White and Volker Hartenstein and Kevin Eliceiri and Pavel Tomancak and Albert Cardona},
+  title = {Fiji: an open-source platform for biological-image analysis},
+  journal = {Nature Methods}
+}
+
