+++ JOSS/paper.md
@@ -24,19 +24,19 @@ bibliography: paper.bib
 
 # Summary
 
-`SpmImage Tycoon` is a cross-platform application designed for fast and effortless organization, analysis, and editing of scanning probe microscopy images and spectra. The application supports automatic channel-of-interest detection, image editing (such as background corrections and contrast adjustment), keywording, star-ratings, powerful search and filtering (e.g., by keywords, data type, location, scan size), as well as export of the data into a OpenDocument Presentation file format [@oasisodp]. Such easy-to-use tools can lower the entry barrier for aspiring scientists, boost the efficiency of experienced researchers, and help to create and leverage large scanning probe microscopy datasets for machine learning and artificial intelligence applications.
+`SpmImage Tycoon` is a cross-platform application designed for fast and effortless organization, analysis, and editing of scanning probe microscopy images and spectra. The application supports automatic channel-of-interest detection, image editing (such as background corrections and contrast adjustment), keywording, star-ratings, powerful search and filtering (e.g., by keywords, data type, location, scan size), as well as export of the data into an OpenDocument Presentation file format [@oasisodp]. Such easy-to-use tools can lower the entry barrier for aspiring scientists, boost the efficiency of experienced researchers, and help to create and leverage large scanning probe microscopy datasets for machine learning and artificial intelligence applications.
 
 # Statement of need
 
-The evolution of the field of scanning probe microscopy has been associated with development of instrumental capabilities and methods for data acquisition and analysis [@wsxm; @gwyddion; @spectrafox]. However, much less emphasis has been placed on improving the management and organization of the measured datasets. While datasets can consist of thousands of images and spectra per project, it is not uncommon for researchers to go through each image or spectrum and manually select the channels of interest for each one – only to repeat the same procedure every time they want to analyze their data again. `SpmImage Tycoon` aims to solve this problem by providing a fast and convenient way to organize, analyze, and edit images and spectra obtained in scanning tunneling microscopy (STM) and atomic force microscopy (AFM) experiments.
+The evolution of the field of scanning probe microscopy has been associated with development of instrumental capabilities and methods for data acquisition and analysis [@wsxm; @gwyddion; @imagej; @imagej2; @fiji; @spectrafox]. However, much less emphasis has been placed on improving the management and organization of the measured datasets. Thus, a typical workflow for researchers is to use one of the capable tools for processing and analysis of images and spectra (such as WSxM [@wsxm], Gwyddion [@gwyddion], ImageJ [@imagej], ImageJ2 [@imagej2], Fiji [@fiji], and SpectraFox [@spectrafox]), but filesystem folders and manually created lists of best measurements for organization of the data.  While datasets can consist of thousands of images and spectra per project, it is not uncommon for researchers to go through each image or spectrum and manually select the channels of interest for each one – only to repeat the same procedure every time they want to analyze their data again. `SpmImage Tycoon` aims to solve this problem by providing a fast and convenient way to organize, analyze, and edit images and spectra obtained in scanning tunneling microscopy (STM) and atomic force microscopy (AFM) experiments.
 
 # Implementation
 
 `SpmImage Tycoon` is written in Julia [@Julia-2017] and uses JavaScript/Electron [@electronjs] for the frontend. It can be easily installed and updated via Julia's package manager [@juliapkg] and runs on all major desktop platforms.
 
[…]
