### 2022-08-24 #4644 @alexriss
To address your comments, I have now updated the paper and the project README.md, on github. Furthermore, I added a CODE_OF_CONDUCT.md and CONTRIBUTING.md.

A point-by-point response to your comments follows.

Regarding the project documentation:
- I expanded the feature list in the project documentation, see https://github.com/alexriss/SpmImageTycoon.jl#features (it is more detailed than in the paper)
- Added more detailed description of the files and how to contribute to the project.
- As outlined above, no particular changes on how to run the tests - but I added a link to the test dataset.
- As outlined above, there is no explicit documentation of the API in this project - but I have added a description of where to find the files for the HTML/CSS/JavaScript frontend, as well as for the Julia backend.

Regarding the paper:
- I have expanded the "statement of need" and now list other processing/analysis tools and typical data organization workflows - which I hope this app improves.
- corrected the typos - thank you!
- more details on the improvement of the workflows are given (see above); reformulated the sentence on addition of new features (more explanation is given below).
- I would prefer to not add a screenshot to the paper. The app is still being developed and I plan to add a menu and improve the navigation more, and I would like to avoid having outdated screenshots in a paper that will be available for many years to come. Of course there are screenshots and videos in the project documentation - potential users will hopefully always look at the project website - and I can keep these up-to-date.

Regarding contributions from other people, I have added a few sections to the project documentation. However, it is not completely easy: there is a Julia backend and a HTML/CSS/JavaScript (via Electron) frontend. And of course these need to communicate. So for most new features, coders will need to adapt both, the backend and the frontend. And for this, contributors will have to "dive" into the code to some extent. The best starting point is probably to get in touch with me.

However, for image or spectrum filters, I am planning to add an interface, in which users can stack mathematical operations (e.g. Laplace, Gauss filters). This would be very flexible and hopefully accessible to end users. And there could even be a plugin interface. But this is more of a long-term project...

Thanks again for all your efforts. I am very happy to hear that you found the software easy to use - this was one of my main design goals!
