### 2022-08-24 alexriss/SpmImageTycoon.jl#11 @kasasxav
It worked !
