### 2022-08-24 #4644 @kasasxav
Hi!

Great! It looks good for me. I have checked the review points that were missing.

Just check that the new sentence you added in the paper is not gramatically complete: "but filesystem folders and manually created lists of best measurements for organization of the data" needs a verb.

I also think you can expand it a bit more. From the cited works I have only worked with ImageJ. I think there is a clear difference between SpmImage Tycoon and ImageJ. While ImageJ focuses on the processing and analysis of the images, tycoon is more about the organization of them, the accessibility of the metadata and the fact that you can explore easily within a folder and correct multiple images together, for example. So I would say expanding on that would definitely be interesting. For example, if you have a folder with 100 images and you need to browse through them and apply some basic operations easily, intuitively and quickly, I think that would become challenging with ImageJ. And most likely you find similar reasons for the other software packages, I think that would be helpful.

I think the image filtering idea would be really nice! You could even have a generic function that people can reimplement with the desired operation, with an input and output image. I don't know if it could be some kind of "superclass" that then you can create "subclasses" of it, so that all the links with the backend are already implemented in the "superclass" (sorry for my lack of Julia knowledge). I don't see this as a requirement for the revision, though.

I also added an issue that @alexriss  solved: https://github.com/alexriss/SpmImageTycoon.jl/issues/11
