### 2022-08-25 alexriss/SpmImageTycoon.jl#11 @alexriss
Great, thanks for the help. And yes, it was in _my_ code....  😃
