### 2022-08-26 #4644 @alexriss
Hi - thanks again. Yes, exactly, these apps are focused on processing and analysis, but they aren't really geared as much towards browsing and organization - and with that batch editing of course gets harder. I have seen plenty of scanning probe users' workflows and many of them really just create lists of their best data, or copy the best files into some directory, or they keep browsing through all images every time to find the best images/spectra. I wanted to improve the efficiency of these workflows, this is why I started this project in the first place. And I think this is now more clearly outlined in the paper: the other apps are great for editing/analysis, but tycoon's focus is more on organization and management - with some editing (also batch editing) and analysis features baked in, and more to come - hopefully with a new type of filter or plugin system at some point.
