+++ JOSS/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 # Statement of need
 
-The evolution of the field of scanning probe microscopy has been associated with development of instrumental capabilities and methods for data acquisition and analysis [@wsxm; @gwyddion; @imagej; @imagej2; @fiji; @spectrafox]. However, much less emphasis has been placed on improving the management and organization of the measured datasets. Mature and capable tools for processing and analysis of images and spectra exist, e.g., WSxM [@wsxm], Gwyddion [@gwyddion], ImageJ [@imagej], ImageJ2 [@imagej2], Fiji [@fiji], and SpectraFox [@spectrafox]. However, the workflow for browsing and manage data is less developed and thus users often resort to organization in filesystem folders and manually created lists of the best measurements. This can make common operations of tagging and editing selected subsets of images or spectra rather cumbersome. Furthermore, while datasets can consist of thousands of images and spectra per project, it is not uncommon for researchers to go through each image or spectrum and manually select the channels of interest for each one – only to repeat the same procedure every time they want to analyze their data again. `SpmImage Tycoon` aims to solve these problems by providing a fast and convenient way to browse, organize, analyze, and batch edit images and spectra obtained in scanning tunneling microscopy (STM) and atomic force microscopy (AFM) experiments.
+The evolution of the field of scanning probe microscopy has been associated with development of instrumental capabilities and methods for data acquisition and analysis [@wsxm; @gwyddion; @imagej; @imagej2; @fiji; @spectrafox]. Much less emphasis has been placed on improving the management and organization of the measured datasets. Mature and capable tools for processing and analysis of images and spectra exist, e.g., WSxM [@wsxm], Gwyddion [@gwyddion], ImageJ [@imagej], ImageJ2 [@imagej2], Fiji [@fiji], and SpectraFox [@spectrafox]. However, the workflow for browsing and managing data is less developed and thus users often resort to organization in filesystem folders and manually created lists of the best measurements. This can make common operations of tagging and editing selected subsets of images or spectra rather cumbersome. Furthermore, while datasets can consist of thousands of images and spectra per project, it is not uncommon for researchers to go through each image or spectrum and manually select the channels of interest for each one – only to repeat the same procedure every time they want to analyze their data again. `SpmImage Tycoon` aims to solve these problems by providing a fast and convenient way to browse, organize, analyze, and batch edit images and spectra obtained in scanning tunneling microscopy (STM) and atomic force microscopy (AFM) experiments.
 
 # Implementation
 
