### 2022-08-31 #4644 @jgostick
Alright, great job @kasasxav and @jingpengw! Thanks so much for your time and energy, which I know is a valuable and limited resource.

@alexriss, congrats, I will recommend this paper to be published!  But first there are a few more items to check off:

- [ ] Set article DOI with ``@editorialbot set <DOI here> as archive``
- [ ] Do a 'dry run' of acceptance with ``@editorialbot recommend-accept``
